# docs/OPENAPI.yaml
...
openapi: 3.1.0
info:
  title: UBUNTU-NTU API
  description: Offline-first African language learning. Progress synchronises as
    an append-only event log; server state is derived by deterministic replay.
  version: 0.1.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /v1/contributions/audio:
    post:
      tags:
        - contributions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                dialectId:
                  type: string
                  minLength: 2
                  maxLength: 32
                promptText:
                  type: string
                  minLength: 1
                audioPath:
                  type: string
                  minLength: 1
                vocabItemId:
                  type: string
                  maxLength: 128
                durationMs:
                  type: integer
                  exclusiveMinimum: true
                  minimum: 0
                snrDb:
                  type: number
              required:
                - dialectId
                - promptText
                - audioPath
              additionalProperties: false
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                  dialectId:
                    type: string
                  countryCode:
                    type: string
                required:
                  - id
                  - status
                  - dialectId
                  - countryCode
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - error
                  - message
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - error
                  - message
                additionalProperties: false
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - error
                  - message
                additionalProperties: false
  /v1/contributions/{id}/vote:
    post:
      tags:
        - contributions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vote:
                  type: string
                  enum:
                    - approve
                    - reject
                    - wrong_dialect
                reason:
                  type: string
                  maxLength: 50
              required:
                - vote
              additionalProperties: false
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  recorded:
                    type: boolean
                    enum:
                      - true
                  vote:
                    type: string
                required:
                  - recorded
                  - vote
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - error
                  - message
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - error
                  - message
                additionalProperties: false
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - error
                  - message
                additionalProperties: false
        "409":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - error
                  - message
                additionalProperties: false
  /v1/contributions/queue:
    get:
      tags:
        - contributions
      parameters:
        - schema:
            type: string
            minLength: 2
            maxLength: 32
          in: query
          name: dialectId
          required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  dialectId:
                    type: string
                  countryCode:
                    type: string
                  contributions:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        promptText:
                          type: string
                        audioPath:
                          type: string
                        status:
                          type: string
                      required:
                        - id
                        - promptText
                        - audioPath
                        - status
                      additionalProperties: false
                required:
                  - dialectId
                  - countryCode
                  - contributions
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - error
                  - message
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - error
                  - message
                additionalProperties: false
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - error
                  - message
                additionalProperties: false